refactor(ui): centralize cross-process alerts in AlertHelper
